Move healthcheck directory to fix CI build
