+++ packages/arb-node-core/nodehealth/nodehealth_test.go
@@ -1,4 +1,4 @@
-package main
+package nodehealth
 
 import (
 	"fmt"
